requirements.txt
streamlit
langchain
langchain-community
langchain-groq
fastembed
chromadb==0.4.24
protobuf==3.20.3
numpy<2.0
pysqlite3-binary
